requirements.txt
dash~=2.17.1
dash_mantine_components~=0.14.3
openai~=1.38.0
python-dotenv~=1.0.1
Flask~=3.0.3
dash-bootstrap-components~=1.6.1rc2
